+++ src/Moq/Capture.cs
@@ -28,6 +28,7 @@ namespace Moq
 		/// Assert.Equal("Hello!", parameters.Single());
 		/// </code>
 		/// </example>
+		[Matcher]
 		public static T In<T>(ICollection<T> collection)
 		{
 			var match = new CaptureMatch<T>(collection.Add);
@@ -51,6 +52,7 @@ namespace Moq
 		/// Assert.Equal("Hello!", parameters.Single());
 		/// </code>
 		/// </example>
+		[Matcher]
 		public static T In<T>(IList<T> collection, Expression<Func<T, bool>> predicate)
 		{
 			var match = new CaptureMatch<T>(collection.Add, predicate);
@@ -73,6 +75,7 @@ namespace Moq
 		/// Assert.Equal("Hello!", capturedValue);
 		/// </code>
 		/// </example>
+		[Matcher]
 		public static T With<T>(CaptureMatch<T> match)
 		{
 			return Match.Create(match);