+++ src/Moq/MatcherAttribute.cs
@@ -2,7 +2,6 @@
 // All rights reserved. Licensed under the BSD 3-Clause License; see License.txt.
 
 using System;
-using System.ComponentModel;
 
 namespace Moq
 {
@@ -12,9 +11,10 @@ namespace Moq
 	/// matching rules.
 	/// </summary>
 	/// <remarks>
-	/// <b>This feature has been deprecated in favor of the new 
-	/// and simpler <see cref="Match{T}"/>.
-	/// </b>
+	/// <para>
+	/// This attribute may be used in conjunction with <see cref="Match.Create{T}"/>,
+	/// or with the older mechanism described below.
+	/// </para>
 	/// <para>
 	/// The argument matching is used to determine whether a concrete 
 	/// invocation in the mock matches a given setup. This 
@@ -119,9 +119,7 @@ namespace Moq
 	/// // use mock, invoke Save, and have the matcher filter.
 	/// </code>
 	/// </example>
-	[AttributeUsage(AttributeTargets.Method, Inherited = true)]
-	[EditorBrowsable(EditorBrowsableState.Never)]
-	[Obsolete("This feature has been deprecated in favor of `Match.Create`.")]
+	[AttributeUsage(AttributeTargets.Method | AttributeTargets.Property, Inherited = true)]
 	public sealed class MatcherAttribute : Attribute
 	{
 	}