+++ tests/Moq.Tests/Linq/SupportedQuerying.cs
@@ -238,6 +238,7 @@ namespace Moq.Tests.Linq
 				Assert.Equal("foo", foo.Do(5));
 			}
 
+			[Matcher]
 			public TValue Any<TValue>()
 			{
 				return Match.Create<TValue>(v => true);