+++ tests/Moq.Tests/MockFixture.cs
@@ -707,6 +707,7 @@ namespace Moq.Tests
 			Assert.Equal(3, mock.Object.Echo(3));
 		}
 
+		[Matcher]
 		private int IsMultipleOf(int value)
 		{
 			return Match.Create<int>(i => i % value == 0);